Fixes/fix robotic arm model transforms + implementation of new robotic models (#54)
* refactor: fixed and refactored path, plots, and small QoL bugs in preperation for release (first version)

* fixes: plots now optionally follow, and can be scaled according to user choice

* fixes

* fixes: correct added postbuild copmmand for HDF5 dlls, along with 3rd party licenses in new release

* update dev inline with Main branch (#42)

* Update citation link in README.md

* Update author information and release date (#41)

* alpha release v0.1.0 (#40)

* refactor: fixed and refactored path, plots, and small QoL bugs in preperation for release (first version)

* fixes: plots now optionally follow, and can be scaled according to user choice

* fixes

* fixes: correct added postbuild copmmand for HDF5 dlls, along with 3rd party licenses in new release

* Remove badge section from README.md

Removed badge section from README.

* fixes

* Feature/joint friction remodel (#44)

* feat: methods to compute linear velocity and angular velocity jacobians

* WIP: addition of jacobians, and replacement of glm with Eigen in mathematically dense compuations

* refactor: move eigen-based mathematical computations for non-visual computational functions.

* feat: added mass matrix logic to M(joint, link)

* fixes: physics computations surrounding model-numeric mistmatch fixes

* feat: Coriolis Model added

NOTE:
- Does not work properly, may crash program due to high output of computations

* fixes: corrected logic and pipeline for inertia and coriolis computations

NOTES:
- Replace computeI_eff with computeSingleIeff for modular, per-joint inertia calculation.
- Rewrite computeCoriolisDiagonal to use finite differences and new inertia method.
- Simplify deriv() to compute inertia and Coriolis terms internally from state.
- Improve computeJointInertiaContribution to use both translational and rotational effects with joint axis in world frame.
- Update simulation and integration logic to u
- I did a final check with github copilot as this is quite abstract for me

* Refactor/robot metrics (#46)

* refactor: update metrics output to suit report question, and limit file size of each run.

* feat: added save command for log file naming and saving instead of always outputting.

TODO:
- Implement gravity torque into current joint-space formula.
- Add different arms to showcase different abilities.
- Add a general script that works for multiple arms.
- Fix RotateX commands (not essential).
- Get save working for specific file names (not essential).
- Use either python or MATLAB to generate plots for the run data.

NOTES:
- I am using various resources to improve this, new ground for me.

* refactor: undid ALOT of thread work, not necessary and way too time consuming.

READY FOR DATA ANALYSIS

* Update Sim_Engine/EngineLib/include/Robots/RobotSystem.h


* Update Sim_Engine/EngineLib/include/Robots/RobotModel.h


* Update MathLib/MathLib/src/integrators/numerical_integrators.cpp


* Update Sim_Engine/EngineLib/include/Interpreter/Commands/SaveCmd.h


* Update Sim_Engine/EngineLib/include/Interpreter/Commands/SaveCmd.h


* Update Sim_Engine/EngineLib/src/Interpreter/Parser.cpp


* Update Sim_Engine/EngineLib/src/Numerics/IntegrationService.cpp


---------


* feat: implemented basic computation of a joints gravity torque,

NOTES:
- Removed the hardcoded setting of tau_i to 0

* fixes: fixed computational overestimation of wrist gravity in ComputeGravityTorque()

* fixes: fixed gravity inconsistenties with robot root frame, and recomputation of FK

NOTES:
- This is confusing, but also not, its very much not in my comfort zone

* fixes: Inverse dynamics style feedforward corrected to be inline with the physically correct model.

* feat: added a very basic selector for gravity values

* fixes

* fixes

* merge fixes

* merge fixes

* merge fixes

* merge fixes

* feat: refactor robotsystem and loading to allow other robotic models.

* fixes to robot loading

* Update README with installation and tools information

Added installation instructions and third-party tools used.

* Revise disclaimer in README for clarity

Updated disclaimer for clarity and grammar.

* Remove libraries used section from README

Removed the list of libraries used from the README.

* Remove latest update section from README

Removed outdated update section from README.

* Fixes/fix robotic arm model transforms (#53)

* feat: added two new robot arm models "panda" and "iiwa14"

TODO:
- Add last one "ur5e"
- Add default pos (loaded pos) to json

* feat: added ur5e robot model and specific DSL scripts for testing and evaluation

* fixes: fixed Panda robot model and specific DSL scripts for testing and evaluation, along  with specific damping and friction values

* feat: added iiwa robot model and specific DSL scripts for testing and evaluation

TODO:
- Figure out why iiwa14 model doesnt rotate correctly, and fix existing kinematic updaters to visually work with different model types.
- ADD MATERIALS
- Already started, but beging report.

* feat: added similar scripts for Z1 robotic arm model, more realisitc comparisons with the same integrators!

TODO:
- Add a GENERAL script, acting as a constant.

* feat: extra scripts

* Remove: old scripts removed from repo

* fixes: iiwa14 fixes, added intertial info

TODO:
- Add colours!!!!!!

* fixes: Updated robotic loading and visual kineamtics with robust logic for different robot joint types, and general safety measures

* fixes

* fixes

* feat: add materials per robotic arm links (Colour only for now), move non-visual var types from float to double

TODO:
-Update calls of said methods using floats for what now is double type params.

* fixes: robot visual fk for mixed mesh frame convention

NOTES (IMPORTANT):
- Some robot assets (e.g. Z1 CAD-derived meshes) encode their link frames directly in mesh local space, while others (URDF-style, e.g. Panda) expect FK to be applied at the object level
- Engine-aligned models: bake FK directly into mesh local transform
- URDF-style models: apply FK + visual origin to object transform
- MAY REVISE LATER

* feat: add UR5e integrator stress test script, along with basic colours for each arm

---------

+++ Sim_Engine/EngineLib/src/Interpreter/Commands/RotateJointByCmd.cpp
@@ -47,7 +47,7 @@ namespace commands {
 			}
 
 			// Get starting angle
-			float theta0 = 0.0f;
+			double theta0 = 0.0f;
 			if (!robot->tryGetJointAngleRad(_link, theta0)) {
 				markFailed("rotateJointBy: joint not found (angle)."); 
 				D_FAIL("rotateJointBy: joint not found (angle) for '%s'", _link.c_str());
@@ -88,8 +88,8 @@ namespace commands {
 		const double tolOmegaRad = degToRad(0.5);	// 0.20 deg/s
 		const double settleSec = 0.15;				// must be stable for 100ms - i need to tune this more
 
-		float theta = 0.0f;
-		float omega = 0.0f;
+		double theta = 0.0f;
+		double omega = 0.0f;
 
 		const bool gotTheta = robot->tryGetJointAngleRad(_link, theta);
 		const bool gotOmega = robot->tryGetJointOmegaRad(_link, omega);
@@ -113,7 +113,7 @@ namespace commands {
 			_noProgressT += dt; 
 		}
 
-		const bool posOK = robot->isJointAtTargetRad(_link, (float)tolPosRad); // consider at target if within position tolerance
+		const bool posOK = robot->isJointAtTargetRad(_link, tolPosRad); // consider at target if within position tolerance
 		const bool omegaOK = (absOm <= tolOmegaRad); // consider stopped if omega is small enough
 
 		if (posOK && omegaOK) {
